Fixed table to store the LIST response
