+++ alicloud/plugin.go
@@ -17,6 +17,7 @@ func Plugin(ctx context.Context) *plugin.Plugin {
 			Schema:      ConfigSchema,
 		},
 		TableMap: map[string]*plugin.Table{
+			"alicloud_cas_certificate":         tableAlicloudUserCertificate(ctx),
 			"alicloud_ecs_autoscaling_group":    tableAlicloudEcsAutoscalingGroup(ctx),
 			"alicloud_ecs_disk":                 tableAlicloudEcsDisk(ctx),
 			"alicloud_ecs_image":                tableAlicloudEcsImage(ctx),
@@ -42,7 +43,6 @@ func Plugin(ctx context.Context) *plugin.Plugin {
 			"alicloud_vpc_vpn_customer_gateway": tableAlicloudVpcVpnCustomerGateway(ctx),
 			"alicloud_vpc_vpn_gateway":          tableAlicloudVpcVpnGateway(ctx),
 			"alicloud_vpc_vswitch":              tableAlicloudVpcVSwitch(ctx),
-			"alicloud_user_certificate":         tableAlicloudUserCertificate(ctx),
 		},
 	}
 	return p
